Tableau.Cloud.Migration.Accelerator/Tableau.Cloud.Migration.Accelerator/scripts/requirements.txt
cffi
clr-loader
configparser
DateTime
glob2
numpy
openpyxl
pandas
pycparser
pythonnet
python-dotenv
requests
tableau-tools
tableau_migration
tableauserverclient
typing_extensions
urllib3